requirements.txt
streamlit>=1.32.0
llama-index>=0.12.52
llama-index-llms-gemini
llama-index-embeddings-gemini
google-generativeai>=0.3.2
PyMuPDF>=1.23.6
python-docx>=1.1.0
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
